feat(ui): enhance landing page with scroll hint and restructure content
- Add animated scroll hint button with chevron icon to landing page
- Add bounce-gentle CSS animation for smooth scroll indicator
- Remove welcome page and consolidate content into landing-features component
- Update main page structure to use modular landing-features approach
- Improve user experience with visual scroll prompt and cleaner content organization
