+++ app/page.tsx
@@ -1,5 +1,11 @@
 import { Landing } from "./landing";
+import { LandingFeatures } from "./landing-features";
 
 export default function Home() {
-  return <Landing />;
+  return (
+    <>
+      <Landing />
+      <LandingFeatures />
+    </>
+  );
 }
